+++ packages/yourloops/test/integration/mock/complete-daily-view-data.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025, Diabeloop
+ * Copyright (c) 2024-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -38,7 +38,7 @@ import {
   Prescriptor,
   PumpManufacturer,
   Source,
-  Unit,
+  Unit, Wizard,
   WizardInputMealFat,
   WizardInputMealSource
 } from 'medical-domain'
@@ -111,6 +111,7 @@ import {
   WIZARD_POSITIVE_OVERRIDE_ID,
   WIZARD_UMM_ID,
   WIZARD_UNDELIVERED_ID,
+  WIZARD_WITHOUT_BOLUS_ID,
   ZEN_MODE_ID,
   ZEN_MODE_ID_WITH_GLY
 } from './data.api.mock'
@@ -193,6 +194,31 @@ export const getCompleteDailyViewDataDblg2 = (deviceName: DeviceSystem = DeviceS
     isoWeekday: WeekDays.Friday,
     source: Source.Diabeloop,
   }]
+
+  const mealWithoutBolus: Wizard = {
+      "epoch": 1659984400000,
+      "displayOffset": -120,
+      "normalTime": "2022-08-08T19:10:00.000Z",
+      "timezone": "Europe/Paris",
+      "guessedTimezone": false,
+      "id": WIZARD_WITHOUT_BOLUS_ID,
+      "type": DatumType.Wizard,
+      "source": Source.Diabeloop,
+      "bolusId": "",
+      "bolusIds": new Set([]),
+      "carbInput": 30,
+      "units": "mmol/L",
+      "bolus": null,
+      "bolusPart2": null,
+      "inputTime": "2022-08-08T22:45:00Z",
+      "recommended": {
+        "carb": 0,
+        "correction": 0,
+        "net": 0
+      },
+      "isoWeekday": WeekDays.Thursday
+    }
+    data.wizards = [...data.wizards, mealWithoutBolus]
 
   return completeData
 }
@@ -1718,7 +1744,7 @@ export const getCompleteDailyViewData = (deviceName: DeviceSystem = DeviceSystem
             "net": 10.07
           },
           "isoWeekday": WeekDays.Thursday
-        },
+        }
       ],
       zenModes: [
         {